use bufio
