+++ main.go
@@ -15,7 +15,7 @@ import (
 var typeFunc = make([]*filter.TypeFunc, 0)
 var contentFunc = make([]filter.ContentOption, 0)
 var r = render.NewRenderer(theme.DefaultTheme, theme.DefaultInfoTheme)
-var p printer.Printer = printer.FitTerminal{}
+var p printer.Printer = printer.NewFitTerminal()
 
 const version = "v0.2.1"
 
@@ -68,8 +68,8 @@ func main() {
 				Action: func(context *cli.Context, b bool) error {
 					if b {
 						contentFunc = append(contentFunc, filter.EnableFileMode(r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -82,8 +82,8 @@ func main() {
 				Action: func(context *cli.Context, b bool) error {
 					if b {
 						contentFunc = append(contentFunc, filter.EnableSize(filter.Auto, r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -96,8 +96,8 @@ func main() {
 				Action: func(context *cli.Context, b bool) error {
 					if b {
 						contentFunc = append(contentFunc, filter.EnableOwner(r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -110,8 +110,8 @@ func main() {
 				Action: func(context *cli.Context, b bool) error {
 					if b {
 						contentFunc = append(contentFunc, filter.EnableGroup(r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -124,8 +124,8 @@ func main() {
 				Action: func(context *cli.Context, b bool) error {
 					if b {
 						contentFunc = append(contentFunc, filter.EnableTime("02.Jan'06 15:04", r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -142,8 +142,8 @@ func main() {
 				Usage:   "print by line",
 				Action: func(context *cli.Context, b bool) error {
 					if b {
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -177,8 +177,8 @@ func main() {
 						}
 						typeFunc = newFF
 						contentFunc = append(contentFunc, filter.EnableFileMode(r), filter.EnableSize(filter.Auto, r), filter.EnableOwner(r), filter.EnableGroup(r), filter.EnableTime("06 Jan 02 15:04", r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
@@ -191,8 +191,8 @@ func main() {
 				Action: func(context *cli.Context, b bool) error {
 					if b {
 						contentFunc = append(contentFunc, filter.EnableFileMode(r), filter.EnableSize(filter.Auto, r), filter.EnableOwner(r), filter.EnableGroup(r), filter.EnableTime("06 Jan 02 15:04", r))
-						if _, ok := p.(printer.Byline); !ok {
-							p = printer.Byline{}
+						if _, ok := p.(*printer.Byline); !ok {
+							p = printer.NewByline()
 						}
 					}
 					return nil
